fix: 새로운 파티 신청 - 다음 버튼 컴포넌트 구조 수정
