+++ src/pages/NewPartyPage/MemberAndDate/index.jsx
@@ -1,14 +1,18 @@
-import { useEffect } from "react";
-import PartyDate from "./PartyDate";
 import Member from "./Member";
+import PartyDate from "./PartyDate";
+import PageButton from "./PageButton";
 
-function MemberAndDate({ member, setMember, date, setDate, setNextOK }) {
-  useEffect(() => setNextOK(true));
-
+function MemberAndDate({ region, driverId, member, setMember, date, setDate }) {
   return (
     <>
       <Member member={member} setMember={setMember} />
       <PartyDate date={date} setDate={setDate} />
+      <PageButton
+        region={region}
+        member={member}
+        date={date}
+        driverId={driverId}
+      />
     </>
   );
 }
